requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
docstring_parser==0.17.0
dotenv==0.9.9
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-api-core
google-auth==2.43.0
google-cloud-aiplatform
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager
google-cloud-storage==2.19.0
google-cloud-vision
google-crc32c==1.7.1
google-genai==1.53.0
google-resumable-media==2.8.0
googleapis-common-protos
grpc-google-iam-v1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
proto-plus
protobuf
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.52.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vertexai==1.71.1
websockets==15.0.1
torch
pix2tex[gui]
nougat-ocr
